rework system naming convention
for the time being, it seems as if we do not actually
care about the letter at the end of the model number,
as it is only used to denote SoC changes. since the
SoC in a machine does not (yet) have any bearing on the
physical audio device in the corresponding machine, we
can ignore it.

this lets us avoid duplication by however many SoC
variants exist of a particular physical machine.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>