j415: rework compressors
change the compressor architecture such that we reduce reliance on
the multiband compressor, and add a fast, hard single band at the
output of the woofer IRs.

this serves the dual purpose of reducing the CPU use of the multiband
by about half, and increasing the dynamic range of the array by
waiting until after the IRs to deal with bass and sub-bass clipping.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>