j314: fix speakersafetyd panic
On certain inputs with significant output around certain frequencies,
we suffer from slight overexcursion at full volume. Speakersafetyd panics
as the woofers get hot and the voltage and current readings become odd when
the cone is unable to move freely. Tone down the IRs slightly and extend
the compressor bandpass to capture the frequencies where this happens.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>