j314: rework IRs and bankstown parameters
previously, there was a noticeable "dead zone" effect even when dead centre
with the machine. redo the IRs for j314 from scratch to deal with these. at
the same time, rework the bankstown parameters to produce a pleasant sound
with bankstown 1.1.0.

this results in a much more linear and balanced sound than we had before

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>