firs: add device.icon-name prop to catpure.props
Unlike Plasma, GNOME does not have a fallback device icon for
sinks and sources where it is not explicitly specified by the sound server.
Add an explicit icon to the user-visible sink so that GNOME has a nice
speaker icon next to the sink name in its mixer.

This will also apply to Plasma once it gets an audio-speaker icon, which will
look much better than the fallback audio-card icon.

Reported-by: Jonas Dreßler <verdre@v0yd.nl>
Signed-off-by: James Calligeros <jcalligeros99@gmail.com>