conf: port volume limiter to WP 0.5 config style
WirePlumber 0.5 has moved to a new paradigm for calling scripts and
configuring itself that eliminates the use of Lua for configuration.

To call our volume-limiting script now, we must mark it as a custom
Lua extension in the WirePlumber config fragment we ship and enable
it. We must also move it from main.lua.d, as this directory no longer
exists.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>