conf: wireplumber: match on ALSA path
In the move from Lua to SPA-JSON configuration, the matching rule logic
was changed from AND to OR. Coupled with the fact that UCM nodes are now
renamed to their UCM profile name, this meant that multiple rules would match
for any given device and trigger multiple virtual nodes.

Use only the ALSA path for the speaker output subdevice to match our rules.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>