conf: wireplumber: Fix & force speakers node name
PipeWire commit 833c86d35 changed the profile name and therefore the
node name for the internal speakers device. The only reason it was still
working is because fallback device selection usually picked the internal
speakers device.

To fix this and avoid the problem in the future, force the node name to
a constant in the WP config and then target that in the DSP configs.
Also set dont-fallback so that if this fails for some reason, the DSP
doesn't try to bind to some other random device.

While we're here also make the nick and description more helpful (this
mostly affects how the node shows up in Carla/JACK, since it's otherwise
hidden... hopefully).

Signed-off-by: Asahi Lina <lina@asahilina.net>