conf: bump uclamp.max to 128
Changes to task placement and scheduling in Linux 6.13 has led
to the ecores being (correctly) utilised for more things.
However, because we are telling the scheduler to effectively
ignore PipeWire and WirePlumber when making capacity decisions with
such a low uclamp.max, this leads to the ecores choking. This manifests
as xruns when trying to play software-decoded videos. Which is currently
all videos...

Bump uclamp.max to 128 so that audio plays smoothly when decoding videos.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>