speakers: name DSP graph output node to avoid confusion
Patchbays and DAWs, like qpwgraph and Carla, will use the full filter-chain's
name if no node-specific name is given. This can be confusing. Let's give the
DSP output its own name so that it is more obvious that it shouldn't be used
for anything.

Unfortunately we cannot do the same with the input as that will overwrite the
friendly name we actually want the sink to have.

Signed-off-by: James Calligeros <jcalligeros99@gmail.com>